Make new base class for "real" integraiton test
Rather than try to shoe horn integration tests into one or two subclasses
just make a new one and insert it into the class hierarchy. The whole point
of this is so that it is obvious from NamingConventionCheck's perspective
what the name of a class should be without making weird changes to the
application.
